+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Implement target detailspage queries and mutations [#2822](https://github.com/greenbone/gsa/pull/2822)
 - Refactor target component to use useReducer [#2812](https://github.com/greenbone/gsa/pull/2812)
 - Implement nvt listpage queries and mutations [#2798](https://github.com/greenbone/gsa/pull/2798)
+- Use graphql on override detailspage [#2776](https://github.com/greenbone/gsa/pull/2776)
 - Implement nvt detailspage queries and mutations [#2774](https://github.com/greenbone/gsa/pull/2774)
 - Applied Backend changed to scanconfigs and policies [#2783](https://github.com/greenbone/gsa/pull/2783)
 - Refactor credential dialog into function component [#2772](https://github.com/greenbone/gsa/pull/2772)